% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_extinct.R
\name{B_extinct}
\alias{B_extinct}
\title{Set abundances to 0 when B goes extinct}
\usage{
B_extinct(N)
}
\arguments{
\item{N}{Community abundance matrix with \code{nrow = n_sp = 3} and \code{ncol = n_patch}}
}
\value{
Abundance matrix, N.
}
\description{
Set abundances to 0 when B goes extinct
}
\details{
This is an internal function which sets higher trophic levels (species C and P), to 0 in a patch when the abundance of species B is 0.
}
\examples{
n_sp = 3
n_patch = 5
N <- matrix(rpois(n = n_sp*n_patch, lambda = 100), nrow = n_sp, ncol = n_patch)
N
# make a few Row 1 == 0
N[1, 1:2] <- 0
N
B_extinct(N)

}
