% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_b.R
\name{calc_b}
\alias{calc_b}
\title{calculate parameter b which determines asymptotic level of prey recruitment}
\usage{
calc_b(r_max, k)
}
\arguments{
\item{r_max}{maximum per capita recruitment rate for basal species B}

\item{k}{carrying capacity}
}
\value{
parameter b
}
\description{
calculate parameter b which determines asymptotic level of prey recruitment
}
\examples{
calc_b(r_max = 2, k = 100)

}
