% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mat_internal.R
\name{dist_mat_internal}
\alias{dist_mat_internal}
\title{Internal distance matrix function}
\usage{
dist_mat_internal(dist_mat, landscape_size, n_patch)
}
\arguments{
\item{dist_mat}{Distance matrix describing the spatial arrangement of patches. should be a square matrix with \code{ncol = nrow = n_patch} and \code{diagonal = 0}. If \code{dist_mat = NULL}, a square landscape with dimension \code{landscape_size} and \code{n_patch} randomly distributed will be calculated.}

\item{landscape_size}{Size of each side of randomly generated 2D habitat. Only used if \code{dist_mat = NULL}}

\item{n_patch}{the number of patches, inherited from the \code{igp_sim()} function argument.}
}
\value{
Checks that supplied distance matrix is in correct format and returns it. If no matrix is supplied, it calculates one and returns it.
}
\description{
Internal distance matrix function
}
\examples{
n_patch = 5
landscape_size = 10
x_coord = runif(n_patch, 0, landscape_size)
y_coord = runif(n_patch, 0, landscape_size)
dist_mat = data.matrix(dist(cbind(x_coord, y_coord),
                            diag = TRUE, upper = TRUE))

dist_mat_internal(dist_mat = dist_mat, landscape_size = 10, n_patch = n_patch)

}
