% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturb_internal.R
\name{disturb_internal}
\alias{disturb_internal}
\title{Determine if a disturbance occured and reduce population abundances in a time step.}
\usage{
disturb_internal(
  N,
  adjacency_matrix,
  dist_mat,
  disturb_type,
  disturb_p,
  disturb_value,
  disturb_rho,
  disturb_decay,
  river_network_structure
)
}
\arguments{
\item{N}{Abundance matrix with \code{nrow = n_sp = 3} and \code{ncol = n_patch}.}

\item{adjacency_matrix}{Square matrix (\code{n_patch * n_patch}) describing which patches are adjacent to each other. This is used for branching river networks, and is designed to take the \code{adjacency_matrix} output from \code{brnet()} function as an input.}

\item{dist_mat}{Distance matrix indicating the distance between patches. This is designed to take the \code{distance_matrix} output from \code{brnet()} as an input. If it is not supplied (\code{dist_mat = NULL}), \code{n_patch} are assumed to be randomly dispersed in a square, 2D landscape with dimensions \code{landscape_size * landscape_size}}

\item{disturb_type}{Character vector of one of \verb{c("point-source", "regional", or NULL)}. "point-source" applies disturbances randomly to individual patches. If it is a river network, the disturbance travels downstream and decays according to the \code{disturb_decay}  If it is a 2D network, the disturbance decays with increasing distance from disturbed patch, and is controlled by \code{disturb_rho} argument}

\item{disturb_p}{probability of disturbance occurring. If \code{disturb_type = "point-source"}, whether or not a disturbance occurs is determined independently for each patch via \code{rbinom(n = n_patch, size = 1, prob = disturb_p)}. If \code{disturb_type = "regional"}, disturbance for the entire meta-community is determined in each time step via \code{rbinom(n = 1, size = 1, prob = disturb_p)}.}

\item{disturb_value}{Vector of numbers describing an environmental condition in patch \code{i}. This vector will be scaled from 0 to 1 using the inverse logit function. This is designed to take the output of \code{df_patch$environment} from \code{brnet()} as an input.}

\item{disturb_rho}{integer >= 0. controls how quickly disturbance diminish with distance: 0 = no decay, all patches experience equal impact; 10 impact rapidly decays, with adjacent patches hardly being affected.}

\item{disturb_decay}{numeric from 0 to 1. Indicates what percent of the disturbance remains at the next patch (i.e. 0.1, 0.5, 0.75, 1 = 10, 50, 75 and 100\% of disturbance impacts in adjacent patch, respectively.}

\item{disturb_mag}{(\code{deprecated}) numeric value (0 to 1) controlling the initial magnitude of impact given a disturbance occurs. 0.1, 0.75 = population abundances reduced by 10 or 75\%, respectively. Assuming a single disturbance (i.e. adjacent or upstream patches not also disturbed) this values indicates the maximum disturbance magnitude a patch will be subjected to in a single time step.}
}
\value{
List with two elements. \code{N} = abundance matrix after accounting for disturbances. Values don't necessarily have to be integers. Numeric-double values will be converted to integer with \code{rpois()} in the simulation function. \code{patch_extinction} is an integer vector of \code{length = n_patch} indiciating if a disturbance did happen (1) or did not happen (0).
}
\description{
Determine if a disturbance occured and reduce population abundances in a time step.
}
\examples{
disturb_internal(N, adjacency_matrix, dist_mat, environment_value, disturb_type, disturb_p, disturb_mag, disturb_rho, disturb_decay)

}
