% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_to_k.R
\name{env_to_k}
\alias{env_to_k}
\title{Calculate K based on environmental value}
\usage{
env_to_k(env, k_base)
}
\arguments{
\item{env}{Vector of environmental values.}

\item{k_base}{Mean carrying capacity K.}
}
\value{
Vector of carrying capacities
}
\description{
Calculate K based on environmental value
}
\details{
This function re-scales a vector of environmental values for patches to carrying capacities. the minimum carrying capacity = \code{0.5 * k_base} and the maximum carrying capacity = \code{1.5 * k_base}.

This function is designed to take \code{brnet()$df_patch$environment} as the \code{env} argument. See the \code{brnet()} function to control the headwater heterogeneity and spatial auto-correlation of environmental values between patches.
}
\examples{
env_to_k(env = c(0.5, 1, 1.5), k_base = 100)

}
