% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcl_prop.R
\name{fcl_prop}
\alias{fcl_prop}
\title{Title}
\usage{
fcl_prop(fcl)
}
\arguments{
\item{fcl}{Vector of FCL states, output from \code{get_fcl()}}
}
\value{
vector of length 5, with each element being the proportion of patches in that FCL state.
}
\description{
Title
}
\details{
Designed to accept the \code{get_fcl()} output as an input vector. \code{fcl} should only be composed of values = \code{c(0, 1, 2, 2.5, 3)}
}
\examples{
# simulate abundance matrix
# nrow = n_sp = 3
# ncol = n_patch = 10
N <- matrix(rpois(3 * 10, lambda = 1), nrow = 3, ncol = 10)

# make sure abundance of C and P = 0 when abundance of B = 0
N[,N[1,] == 0] <- 0
# calculate food chain length state
fcl_state <- get_fcl(N)
# calculate proportion of patches in a given state
fcl_prop(fcl = fcl_state)


}
