% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igp_sim.R
\name{igp_sim}
\alias{igp_sim}
\title{Simulate IGP dynamics in 3 species communities through space and time}
\usage{
igp_sim(
  n_patch = 20,
  n_0 = NULL,
  dist_mat = NULL,
  landscape_size = 10,
  adjacency_matrix = NULL,
  k_function = NULL,
  n_upstream = NULL,
  environment_value = NULL,
  k_base = 150,
  k_c = 10,
  k_min_exponent = 1.1,
  k_max_exponent = 1.35,
  p_dispersal = 0.1,
  theta = 1,
  r_max = 2.5,
  ebc = 2,
  ebp = 2,
  ecp = 2,
  alphabc = 4,
  alphap = 4,
  betabc = 20,
  betap = 20,
  P_pref = 0.25,
  s0 = 0.75,
  disturb_type = NULL,
  disturb_value = NULL,
  disturb_p = 1e-04,
  disturb_mag_mean = 0.9,
  disturb_mag_sd = 0.1,
  disturb_rho = 1,
  disturb_decay = 0.75,
  n_burnin = 200,
  n_timestep = 1000,
  plot_disturbance = FALSE,
  plot_fcl = FALSE,
  plot_patch_dynamics = FALSE
)
}
\arguments{
\item{n_patch}{single numeric value. Total number of patches in simulation}

\item{n_0}{numeric value length should be either 1 or 3. starting abundances for species in each patch. If only one value is supplied it is assumed to be the same for all three species. Can designate starting values for each species with a vector of length 3. If NULL starting abundances for B, C, and P are 0.8, 0.5, and 0.25 * k_base, respectively.}

\item{dist_mat}{square distance matrix describing distance of all patches. If it is NULL, assumes a square landscape with \code{length = landscape_size}}

\item{landscape_size}{single numeric value. Length of a landscape on a side.}

\item{adjacency_matrix}{square matrix describing which patches are adjacent to one another.}

\item{k_function}{character string, one of: \code{c("patches-upstream", "random", "environment", NULL)}. \code{"patches-upstream"} calculates carrying capacity based on the number of nodes upstream. \code{"random"} assigns an integer value to each patch from a Poisson distribution with \code{lambda = k_base}. \code{"environment"} calculates K based on \code{df_patch$environment} value returned from \code{brnet()} function. \code{NULL} assumes that \code{k = k_base} in all patches.}

\item{n_upstream}{numeric vector of \code{length = n_patch}. used if \code{k_function = "patches-upstream"}}

\item{environment_value}{numeric vector of \code{length = n_patch}. used if \code{k_function = "environment"}}

\item{k_base}{single numeric value. Used to calculate carrying capacity, K. see ?k_function_internal for more.}

\item{k_c}{single numeric value. Constant used to calculate K when \code{k_function = "patches-upstream"}. Default = 10}

\item{k_min_exponent}{single numeric value. Minimum value of a uniform distribution for the exponent used to calculate K when \code{k_function = "patches_upstream"}}

\item{k_max_exponent}{single numeric value. Maximum value of a uniform distribution for the exponent used to calculate K when \code{k_function = "patches_upstream"}}

\item{p_dispersal}{numeric value (length should be one or 3). Probability of dispersal.}

\item{theta}{numeric value of length 1 or 3. Dispersal parameter describing dispersal capability of species. If length(theta) = 1 it is the same for all three species. Can be set for each species with vector of length = 3.}

\item{r_max}{single numeric value. Maximum reproductive number for the basal species, B, of the Beverton-Holt model.}

\item{ebc}{single Numeric value. Conversion efficiency of turning B biomass into new C biomass}

\item{ebp}{single Numeric value. Conversion efficiency of turning B biomass into new P biomass}

\item{ecp}{single Numeric value. Conversion efficiency of turning C biomass into new P biomass}

\item{alphabc}{single numeric value. Parameter controlling the predation of resource B by consumer C}

\item{alphap}{single numeric value. Parameter controlling the predation of both resources by consumer P}

\item{betabc}{single numeric value. Parameter controlling the predation of resource B by consumer C}

\item{betap}{single numeric value. Parameter controlling the predation of both resources by consumer P}

\item{P_pref}{single numeric value between 0-1 or NULL. Predator preference of B over C. if \code{NULL} (maybe NA?) calculates predator preference based on conversion efficiency rates and relative abundances of both prey species, i.e., Holling's "type 3" response.}

\item{s0}{numeric value length should be either 1 or 3. Base survival probability for all (\code{length = 1}) or each (\code{length = 3}) species}

\item{disturb_type}{character string, one of: \verb{c("point-source", "regional", or NULL)}. \code{"point-source"} applies a disturbance event with magnitude = \code{disturb_mag} to each patch with probability = \code{disturb_p} for each time step, and the magnitude of the disturbance decays moving downstream according to \code{disturb_rho} argument. \code{"regional"} applies a disturbance to the entire network in a given time step with probability = \code{disturb_p}. The magnitude of the disturbance varies based on the \code{environment_value}. \code{NULL} does not apply any disturbance events.}

\item{disturb_p}{single numeric value between 0-1. The probability that a disturbance occurs at a given time step.}

\item{disturb_rho}{Single numeric value >= 0. controls how quickly disturbance diminish with distance: 0 = no decay, all patches experience equal impact; 10 impact rapidly decays, with adjacent patches hardly being affected.}

\item{disturb_decay}{Single numeric value from 0 to 1. Indicates what percent of the disturbance remains at the next patch (i.e. 0.1, 0.5, 0.75, 1 = 10, 50, 75 and 100\% of disturbance impacts in adjacent patch, respectively.}

\item{n_burnin}{Single numeric value. The number of time-steps to occur before recording values. default = 200,}

\item{n_timestep}{Single numeric value. The number of time-steps to be saved. default =  1000}

\item{plot_disturbance}{logical. If \code{TRUE} plots all patches that underwent a disturbance event.}

\item{plot_fcl}{logical. If \code{TRUE} plots the proportion of patches with a given food chain length through all time-steps.}

\item{plot_patch_dynamics}{logical. If \code{TRUE} plots population abundances through time for 5 random patches.}

\item{disturb_mag}{single numeric value between 0-1. This controls the initial magnitude of impact given a disturbance occurs. For example, if a disturbance does occur, a value of 0.1, or 0.75 results in population abundances for each species being reduced by 10 or 75\%, respectively. Assuming a single disturbance (i.e. adjacent or upstream patches not also disturbed) this values indicates the maximum disturbance magnitude a patch will be subjected to in a single time step.}
}
\value{
\code{sp_dynamics} a data frame containing simulated IGP community dynamics.

\code{fcl} a data frame containing the proportion of patches that have a given food chain length for each time step.

\code{sim_params} a data frame which contains the parameter values used in the simulation. This data frame is a single row, unless one of the following argument is of length = 3: \code{p_dispersal}, \code{s0}, or \code{theta}, in which case the data frame will be 3 rows, corresponding to species B, C, and P, respectively.
}
\description{
Simulate IGP dynamics in 3 species communities through space and time
}
\examples{
igp_sim()

}
