% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_sim.R
\name{pop_sim}
\alias{pop_sim}
\title{population dynamic function for each time step}
\usage{
pop_sim(
  N,
  P_pref,
  fixed_P_pref,
  alphabc,
  betabc,
  ebc,
  alphap,
  betap,
  ebp,
  ecp,
  v_s0,
  b,
  k,
  r_max
)
}
\arguments{
\item{N}{Abundance matrix at time t}

\item{P_pref}{value for the predator preference of B over C, numeric between 0-1 or NULL}

\item{fixed_P_pref}{Logical indicating if P_pref is a fixed value \code{TRUE}, or if it varies based on prey abundances \code{FALSE}. This is determined internally based on the input to \code{P_pref} in \code{igp_sim()}}

\item{alphabc}{Parameter controlling the predation of resource B by consumer C}

\item{betabc}{Parameter controlling the predation of resource B by consumer C}

\item{alphap}{Parameter controlling the predation of both resources by consumer P}

\item{betap}{Parameter controlling the predation of both resources by consumer P}

\item{v_s0}{vector of base survival probabilities. created internally based on s0 argument in \code{igp_sim()}}

\item{b}{parameter controlling asymptotic level of recruitment, calculated as (r_max - 1) / k}

\item{k}{carrying capacity for a patch.}

\item{r_max}{maximum per capita reproduction rate of basal species B}
}
\value{
Abundance matrix N, at time t+1.
}
\description{
population dynamic function for each time step
}
\details{
This function estimates the continuous effects of predation and survival probability through one time step (i.e. 1 season), and discrete reproduction at the end of the time step.
}
